+++ src/main/java/likelion/running/domain/signUp/SignUpResult.java
@@ -1,4 +1,4 @@
-package likelion.running.domain.login;
+package likelion.running.domain.signUp;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
