+++ src/main/java/likelion/running/web/dto/memberDto/MemberDto.java
@@ -1,14 +1,9 @@
 package likelion.running.web.dto.memberDto;
 
 import lombok.Data;
+
 @Data
 public class MemberDto {
-
-    private String memberId;
-
-    private String passWord;
-
-    private String checkPassWord;
-
-    private String name;
+    Long id;
+    String name;
 }
